backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.1
django-filter==23.5
djangorestframework-simplejwt==5.3.1
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
celery==5.3.6
redis==5.0.3
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
flower==2.0.1
Pillow==10.3.0
requests==2.31.0
beautifulsoup4==4.12.3
whitenoise==6.6.0
gunicorn==21.2.0
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.4.1
